업데이트된 잔액 저장 consumerService/saveAll insert묶어서 처리

+++ src/main/java/com/pro/bankService/controller/request/BankAccountDepositParam.java
@@ -7,6 +7,6 @@ import lombok.Setter;
 @Setter
 public class BankAccountDepositParam {
     private Long account_number;
-    private double balance;
+    private Long balance;
 
 }
